zadatak 2 urađen

+++ Spirala3/promjenaArtikla.php
@@ -15,47 +15,50 @@
       $path = "products/requisites.xml";
     }
     $doc = new DOMDocument();
-    $doc->load($path, LIBXML_NOBLANKS);
-    $doc->formatOutput = true;
-    //da ne bi ostajao prostor koji je cvor zauzimao
-    $doc->preserveWhiteSpace = false;
+    if (file_exists($path)) {
+      $doc->load($path, LIBXML_NOBLANKS);
+      $doc->formatOutput = true;
+      //da ne bi ostajao prostor koji je cvor zauzimao
+      $doc->preserveWhiteSpace = false;
 
-    //ne treba provjera ima li artikla jer on sigurno postoji ako je na stranici
-    $root = $doc->documentElement;
-    $proizvodi = $root->getElementsByTagName("product");
-    $postoji = false;
-    foreach($proizvodi as $proizvod) {
-      $nazivEdituj = $proizvod->getElementsByTagName("name")[0]->nodeValue;
-      if ($nazivEdituj == $_GET["nazivArtiklaEdit"] && $nazivEdituj != $_GET["pNazivArtiklaEdit"]) {
-        //provjera postoji li vec novi naziv
-        $postoji = true;
-        break;
-      }
-    }
-
-    if ($postoji) {
-      $_SESSION["greskaNazivEdit"] = "Proizvod sa unesenim novim nazivom već postoji!";
-    }
-    else {
+      //ne treba provjera ima li artikla jer on sigurno postoji ako je na stranici
+      $root = $doc->documentElement;
+      $proizvodi = $root->getElementsByTagName("product");
+      $postoji = false;
       foreach($proizvodi as $proizvod) {
         $nazivEdituj = $proizvod->getElementsByTagName("name")[0]->nodeValue;
-        if ($nazivEdituj == $_GET["pNazivArtiklaEdit"]) {
-          //editovanje
-          $proizvod->childNodes[1]->nodeValue = $_GET["nazivArtiklaEdit"];
-          $proizvod->childNodes[2]->nodeValue = $_GET["cijenaArtiklaEdit"];
-          $proizvod->childNodes[3]->nodeValue = "img/artikli/" . $_GET["slikaArtiklaEdit"] . ".jpg";
-          unset($_SESSION["greskaNazivEdit"]);
+        if ($nazivEdituj == $_GET["nazivArtiklaEdit"] && $nazivEdituj != $_GET["pNazivArtiklaEdit"]) {
+          //provjera postoji li vec novi naziv
+          $postoji = true;
           break;
         }
       }
+
+      if ($postoji) {
+        $_SESSION["greskaNazivEdit"] = "Proizvod sa unesenim novim nazivom već postoji!";
+      }
+      else {
+        foreach($proizvodi as $proizvod) {
+          $nazivEdituj = $proizvod->getElementsByTagName("name")[0]->nodeValue;
+          if ($nazivEdituj == $_GET["pNazivArtiklaEdit"]) {
+            //editovanje
+            $proizvod->childNodes[1]->nodeValue = $_GET["nazivArtiklaEdit"];
+            $proizvod->childNodes[2]->nodeValue = $_GET["cijenaArtiklaEdit"];
+            $proizvod->childNodes[3]->nodeValue = "img/artikli/" . $_GET["slikaArtiklaEdit"] . ".jpg";
+            unset($_SESSION["greskaNazivEdit"]);
+            break;
+          }
+        }
+      }
+
+      $doc->save($path);
+
+      header("Location: ShopContent.php");
+      die;
     }
-
-    $doc->save($path);
-
-    header("Location: ShopContent.php");
-    die;
-
+    else
+      echo '<script>alert("greska");</scrtipt>';
+      header("Location: ShopContent.php");
+      die;
   }
-  else
-    echo '<script>alert("greska");</scrtipt>';
 ?>
